pyproject.toml
[tool.poetry]
name = "moodlescraper"
version = "0.1.0"
description = ""
authors = ["Alon Barad <Alon710@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
bs4 = "^0.0.1"
python-dotenv = "^1.0.0"
pydantic = "^2.5.3"
dateparser = "^1.2.0"
python-json-logger = "^2.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

